refactor: separated db, controller and routes logic for dashboard
